+++ FrontEnd/src/components/Admin/UserManagement/roles/CreateRole.js
@@ -8,9 +8,8 @@ import { useCustomMutation } from '../../../CommonComponents/ReactQuery/hooks'
 import { keys } from '../../../../model/Constant'
 import { errorMessage, successMessage } from '../../../../tools/toastify'
 
-export default () => {
+export default ({ onClose }) => {
 
-    const [show, setShow] = useState(false)
     const [name, setName] = useState('')
 
     const mutateCreateRole = useCustomMutation(
@@ -18,8 +17,7 @@ export default () => {
         [[keys.ROLES_KEY]],
         () => {
             successMessage('Role Created')
-            setShow(false)
-            setName('')
+            onClose()
         },
         (error) => errorMessage(error?.errorMessage ?? 'Failed')
     )
@@ -35,21 +33,13 @@ export default () => {
 
     return (
         <>
-            <Button className='otjs-button otjs-button-blue' onClick={() => setShow(true)} >New Role</Button>
-            <Modal id='create' show={show} onHide={() => setShow(false)}>
-                <Modal.Header closeButton>
-                    <h2 className='card-title'>Create new role</h2>
-                </Modal.Header>
-                <Modal.Body>
-                    <Row className="align-items-center">
-                        <Form>
-                            <Form.Label>Name*</Form.Label>
-                            <Form.Control type="text" value={name} placeholder='name' onChange={(event) => setName(event.target.value)} required />
-                        </Form>
-                    </Row>
-                    <RoleForm onSubmitRole={create} />
-                </Modal.Body>
-            </Modal>
+            <Row className="align-items-center">
+                <Form>
+                    <Form.Label>Name*</Form.Label>
+                    <Form.Control type="text" value={name} placeholder='name' onChange={(event) => setName(event.target.value)} required />
+                </Form>
+            </Row>
+            <RoleForm onUpdateRole={create} />
         </>
     );
 }
